Avoid script abort when wrong line matched
The search string used to match dependencies and devDependencies is too
broad and captures other lines that contain debug switches like
    ''-- "debug", "log"'
because one of the affected packages is named debug. Configuration lines
like
     "supports-color": { "optional": true
are matched as well leading to script abort. This patch extends the
requirements for a match by including the colon separator and by
preventing the whole script from abortion if no matching version is
found.

If there is no matching version in the line with the package name or the line below (captured by -A1 parameter) the found_version will be an empty string and effectively bypass the check.